refactor: isolate the changes inside of the kilo folders
