+++ packages/opencode/src/kilocode/shell/shell.ts
